requirements.txt
urwid>=2.1.2
python-vlc>=3.0.18121
